WIP: refactor page, decompose
