+++ components/staff_schedule/IMPLEMENTATION.log
@@ -34,3 +34,8 @@
 - renamed bootstrap contract/docs to `months`
 - renamed graph request field/docs to `month`
 - fixed `hours_days` naming in local staff_schedule contract docs
+2026-06-22 18:00 +0300
+- authenticated Figma MCP and located the target design page `График работы`, but further design inspection is blocked by the current Figma MCP seat rate limit
+- adjusted near-term execution plan: keep the module read-only and UI-focused, avoid inventing mutation behavior until backend routes land
+- refactored `staff_schedule` page structure toward project module patterns: extracted header/table sections, added schedule constants, and moved graph/page derivation into a dedicated view-model helper
+- kept current rendering behavior and live/fallback read flow intact so future desktop/mobile parity work can build on a cleaner module shape
