+++ components/staff_schedule/staffScheduleConstants.js
@@ -1,0 +1,18 @@
+export const STAFF_SCHEDULE_SOURCE_MODES = {
+  API: "api",
+  STAFF_SCHEDULE: "staff_schedule",
+  LEGACY: "work_schedule_fallback",
+};
+
+export const EMPTY_PERIOD = {
+  days: [],
+  bonus: [],
+  other_summ: {},
+  order_stat: [],
+  bonus_other: 0,
+};
+
+export const EMPLOYEE_COLUMN_WIDTH = 170;
+export const POSITION_COLUMN_WIDTH = 150;
+export const DAY_COLUMN_WIDTH = 42;
+export const SUMMARY_COLUMN_WIDTH = 76;
